Some improvements

+++ src/main/java/fitness/F1.java
@@ -20,9 +20,9 @@ public class F1 extends FitnessFunction {
     
     
     @Override
-    public double fitness(Individual individual) {
+    public long fitness(Individual individual) {
         // Optimal reached at x = 15. In binary: 11110000...
-        return Math.pow((Utilities.getInstance()).binaryToDecimal(individual) - 15, 2);
+        return (long)Math.pow((Utilities.getInstance()).binaryToDecimal(individual) - 15, 2);
     }
 
 }
